python-c1-tema3-sol-main/requirements.txt
pytest
Flask
requests
pymongo
flask-sqlalchemy
PyJWT
pandas
jsonschema
